refactor: 폴더 구조를 mvc에 맞게 변경
