+++ src/main/java/vendingmachine/Application.java
@@ -1,5 +1,7 @@
 package vendingmachine;
 
+import vendingmachine.controller.VendingMachine;
+
 public class Application {
 	public static void main(String[] args) {
 		VendingMachine vendingMachine = new VendingMachine();
