+++ src/main/java/vendingmachine/model/Beverage.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.model;
 
 import java.util.Objects;
