+++ src/main/java/vendingmachine/model/BeverageShop.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.model;
 
 import java.util.ArrayList;
 import java.util.List;
