+++ src/main/java/vendingmachine/model/Beverages.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.model;
 
 import java.util.List;
