+++ src/main/java/vendingmachine/view/InputUtil.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.view;
 
 import static camp.nextstep.edu.missionutils.Console.*;
