+++ src/test/java/vendingmachine/model/BeverageTest.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.model;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
