+++ src/test/java/vendingmachine/model/MoneyTest.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.model;
 
 import static org.assertj.core.api.Assertions.*;
